requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
cssselect2==0.8.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.111.0
fastapi-cli==0.0.7
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.26.0
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.6
orjson==3.10.18
pandas==2.2.2
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydyf==0.11.0
Pygments==2.19.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-telegram-bot==20.8
pytz==2025.2
PyYAML==6.0.2
redis==5.0.4
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tinycss2==1.4.0
typer==0.16.0
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
weasyprint==61.2
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
